Move functions from mercator_transform.ts to mercator_utils.ts
