+++ src/render/program/sky_program.ts
@@ -3,7 +3,7 @@ import type {Context} from '../../gl/context';
 import type {UniformValues, UniformLocations} from '../uniform_binding';
 import {ITransform} from '../../geo/transform_interface';
 import {Sky} from '../../style/sky';
-import {getMercatorHorizon} from '../../geo/projection/mercator_transform';
+import {getMercatorHorizon} from '../../geo/projection/mercator_utils';
 
 export type SkyUniformsType = {
     'u_sky_color': UniformColor;
